api/pyproject.toml
[tool.poetry]
name = "emcom-cluster"
version = "0.1.0"
description = "App"
authors = ["Jimmy <datamikolaj@protonmail.com>"]
license = "Open surs"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
django-ninja = "^1.1.0"
django-stubs = "^4.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
